#52 - fixed overriding custom properties set by other processes before
